requirements.txt
fastapi[standard]
sqlmodel
psycopg2-binary
python-dotenv
PyJWT
passlib[bcrypt]